refactor: Move decryption logic to Encryption.py
This commit moves the environment variable decryption logic from `app.py` to `Encryption.py` and calls it from `EnvironmentVariables.py`. This resolves the persistent indentation errors and keeps the decryption logic in a more appropriate location.
