+++ apigenerator/resources/1 - Project/1 - BaseProject/src/e_Infra/g_Environment/Encryption.py
@@ -27,3 +27,16 @@ class Encryption:
             modes.GCM(iv, tag),
         ).decryptor()
         return decryptor.update(encrypted) + decryptor.finalize()
+
+def decrypt_environ():
+    if 'CYPHER_TEXT' in os.environ and os.environ['CYPHER_TEXT']:
+        key = bytes.fromhex(os.environ['CYPHER_TEXT'])
+        encryption = Encryption(key)
+        for k, v in list(os.environ.items()):
+            if k not in ['CYPHER_TEXT']:
+                try:
+                    os.environ[k] = encryption.decrypt(v).decode()
+                except:
+                    pass
+        if 'CYPHER_TEXT' in os.environ:
+            del os.environ['CYPHER_TEXT']
